requirements.txt
mss
pyyaml
pyautogui
numpy
opencv-python
colorama